requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-BasicAuth==0.2.0
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-core==1.0.3
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
multidict==4.6.1
phonenumbers==8.10.23
protobuf==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyee==6.0.0
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
redis==3.3.11
requests==2.22.0
rq==1.1.0
rsa==4.0
six==1.13.0
slackclient==2.3.1
slacker==0.13.0
slackeventsapi==2.1.0
twilio==6.33.1
urllib3==1.25.7
Werkzeug==0.16.0
yarl==1.3.0
